requirements.txt
rich
click
spotipy
python-dotenv